Edit set_charger_availability api, expose unlock_charger api

+++ evnex/schema/charge_points.py
@@ -107,8 +107,8 @@ class EvnexElectricityCostSegment(BaseModel):
 
 
 class EvnexChargeProfileSegment(BaseModel):
-    limit: float
-    start: float
+    limit: int
+    start: int
 
 
 class EvnexElectricityCost(BaseModel):
